6537 Gjennomgang av TODOS i prosjektet
Skrev om Knapperad til tsx
