+++ src/felleskomponenter/knapperad/knapperad.tsx
@@ -1,0 +1,42 @@
+import classnames from "classnames";
+import * as Nav from "../../navFrontend";
+
+import "./knapperad.css";
+import { MouseEventHandler } from "react";
+
+interface KnapperadProps {
+  bekreft?: MouseEventHandler<HTMLButtonElement>;
+  bekreftTekst: string;
+  avbryt?: MouseEventHandler<HTMLButtonElement>;
+  avbrytTekst: string;
+  redigerbart: boolean;
+  bekreftRedigerbart?: boolean;
+  spinner?: boolean;
+  size?: "small" | "medium" | "xsmall";
+}
+
+const Knapperad = ({
+  bekreft,
+  bekreftTekst,
+  avbryt,
+  avbrytTekst,
+  redigerbart,
+  bekreftRedigerbart,
+  spinner,
+  size,
+}: KnapperadProps) => {
+  const cls = classnames("container__knapperad");
+
+  return (
+    <div className={cls}>
+      <Nav.Button variant="primary" onClick={bekreft} disabled={!redigerbart || !bekreftRedigerbart} loading={spinner}>
+        {bekreftTekst}
+      </Nav.Button>
+      <Nav.Button variant="tertiary" size={size} onClick={avbryt} disabled={!redigerbart}>
+        {avbrytTekst}
+      </Nav.Button>
+    </div>
+  );
+};
+
+export default Knapperad;
